requirements.txt
streamlit_card
animeworld
universal-startfile